# twitter-go-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance Go backend for the Chanom Twitter clone.",
        "title": "Twitter Go API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/chanombude/twitter-go-api"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of social notifications (likes, follows, etc.) for the current user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "List Notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pagination cursor",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.PageResponse-server_NotificationResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/mark-read": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Clear all unread notification badges for the current user.",
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark All Read",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/stream": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Establish a Server-Sent Events (SSE) connection to receive near real-time notifications for likes, replies, retweets, and new followers.",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Stream Real-time Notifications (SSE)",
                "responses": {
                    "200": {
                        "description": "text/event-stream",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/unread-count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Calculate how many unread notifications the current user has.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Get Unread Count",
                "responses": {
                    "200": {
                        "description": "unread count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperr.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apperr.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                }
            }
        },
...
        "server.NotificationResponse": {
            "type": "object",
            "properties": {
                "actor": {
                    "$ref": "#/definitions/server.UserResponse"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isRead": {
                    "type": "boolean"
                },
                "originalTweetContent": {
                    "type": "string"
                },
                "originalTweetId": {
                    "type": "integer"
                },
                "originalTweetMediaUrl": {
                    "type": "string"
                },
                "tweetContent": {
                    "type": "string"
                },
                "tweetId": {
                    "type": "integer"
                },
                "tweetMediaUrl": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "server.PageResponse-server_NotificationResponse": {
            "type": "object",
            "properties": {
                "hasNext": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.NotificationResponse"
                    }
                },
                "nextCursor": {
                    "type": "string"
                }
            }
        },
...
        "server.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "server.UserResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "bannerUrl": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "followersCount": {
                    "type": "integer"
                },
                "followingCount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isFollowing": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer ' followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}